requirements.txt
streamlit
requests
pillow
opencv-python
numpy
ultralytics
huggingface-hub
